+++ src/dlt_filesystem/source/api.py
@@ -2,6 +2,7 @@ from dlt_filesystem.source.impl.local import LocalFilesystemSource
 from dlt_filesystem.source.impl.remote import (
     AzureSource,
     GCSSource,
+    HDFSSource,
     S3Source,
     SFTPSource,
 )
@@ -9,6 +10,7 @@ from dlt_filesystem.source.impl.remote import (
 __all__ = [
     "AzureSource",
     "GCSSource",
+    "HDFSSource",
     "LocalFilesystemSource",
     "S3Source",
     "SFTPSource",